requirements.txt
gitpython
matplotlib>=3.2.2
numpy>=1.18.5
Pillow>=7.1.2
psutil
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  
torchvision>=0.8.1
tqdm>=4.64.0
torch==1.13.1
torchvision==0.14.1 
packaging==23.0 
pandas==1.5.3 
ipython==8.10.0 
psutil==5.9.4
PyYAML==6.0 
tqdm==4.64.1 
matplotlib==3.7.0 
seaborn>=0.11.0
tensorboard>=2.4.1
